RavenDB-19501 Move BackgroundWorkHelper.cs to \Server\Background
