+++ src/Raven.Server/Smuggler/Documents/SmugglerBase.cs
@@ -12,6 +12,7 @@ using Raven.Client.Documents.Operations.Counters;
 using Raven.Client.Documents.Smuggler;
 using Raven.Client.ServerWide;
 using Raven.Client.Util;
+using Raven.Server.Background;
 using Raven.Server.Documents;
 using Raven.Server.Documents.Expiration;
 using Raven.Server.Documents.Indexes.Auto;
